Embed Discord RPC pipe client in DiscordRpc bot
